requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Flask-Cors==4.0.1
google-cloud-aiplatform==1.55.0